+++ src/main/java/lotto/model/RateCalculator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 public class RateCalculator {
